missioncontrol-exec/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
splunk-sdk==2.0.1
httpx==0.27.0
pydantic==2.8.2
pydantic-settings==2.4.0
